+++ tests/Catalog/CatalogServiceTest.php
@@ -2166,6 +2166,34 @@ final class CatalogServiceTest extends ApiTestCase
         ]));
     }
 
+    public function testGetCompanyByIdWidhMoreInfo(): void
+    {
+        $catalogService = $this->buildCatalogService();
+
+        $company = $catalogService->getCompanyById(3);
+
+        static::assertSame(3, $company->getId());
+        static::assertSame('The World Company Inc.', $company->getName());
+        static::assertSame('the-world-company-inc', $company->getSlug());
+        static::assertSame('The World Company Inc.', $company->getDescription());
+        static::assertSame('40 rue Laure Diebold', $company->getAddress());
+        static::assertSame('01 02 03 04 05', $company->getPhoneNumber());
+        static::assertTrue($company->isProfessional());
+        static::assertEquals(45.778847, $company->getLocation()->getLatitude());
+        static::assertEquals(4.800039, $company->getLocation()->getLongitude());
+        static::assertSame(5, $company->getAverageRating());
+        static::assertSame('Lorem Ipsum', $company->getTerms());
+        static::assertSame('FR', $company->getFullAddress()->getCountry());
+        static::assertSame('40 rue Laure Diebold', $company->getFullAddress()->getAddress());
+        static::assertSame('Paris', $company->getFullAddress()->getCity());
+        static::assertSame('75001', $company->getFullAddress()->getZipCode());
+        static::assertNull($company->getCapital());
+        static::assertNull($company->getLegalStatus());
+        static::assertSame("40483304800023", $company->getSiretNumber());
+        static::assertSame("FR83404833049", $company->getVatNumber());
+        static::assertSame([], $company->getExtra());
+    }
+
     public function testNumericFacet(): void
     {
         $catalogService = $this->buildCatalogService();
@@ -2269,6 +2297,47 @@ final class CatalogServiceTest extends ApiTestCase
         static::assertSame('Puma', $brand->getName());
         static::assertSame('puma', $brand->getSlug());
         static::assertSame(13, $brand->getImage()->getId());
+    }
+
+    public function testGetProductByWithMaxPriceAdjustment()
+    {
+        $catalogService = $this->buildCatalogService();
+        $product1 = $catalogService->getProductById((string) 1);
+        static::assertSame((string) 1, $product1->getId());
+        static::assertSame(24, $product1->getMaxPriceAdjustment());
+
+        $product2 = $catalogService->getProductById((string) 4);
+        static::assertSame((string) 4, $product2->getId());
+        static::assertSame(35, $product2->getMaxPriceAdjustment());
+
+        $product3 = $catalogService->getProductById((string) 3);
+        static::assertSame((string) 3, $product3->getId());
+        static::assertNull($product3->getMaxPriceAdjustment());
+    }
+
+    public function testGetProductByFiltersWithMaxPriceAdjustment(): void
+    {
+        $catalogService = $this->buildCatalogService();
+
+        $filters = (new ProductFilter())->setIds([1, 4, 3]);
+
+        $products = $catalogService->getProductsByFilters($filters);
+        static::assertCount(3, $products);
+
+        $expected = ['1', '3', '4'];
+
+        foreach ($products as $key => $product) {
+            static::assertSame($expected[$key], $product->getId());
+        }
+
+        static::assertSame('1', $products[0]->getId());
+        static::assertSame(24, $products[0]->getMaxPriceAdjustment());
+
+        static::assertSame('4', $products[2]->getId());
+        static::assertSame(35, $products[2]->getMaxPriceAdjustment());
+
+        static::assertSame('3', $products[1]->getId());
+        static::assertNull($products[1]->getMaxPriceAdjustment());
     }
 
     public function testGetProductsWithSubscription(): void